+++ src/components/Overlay/ControlOverlay/WMSOverlay/index.css
@@ -1,0 +1,7 @@
+.loading-spinner {
+    z-index: 1000;
+    position: "absolute";
+    left: "50%";
+    top: "50%";
+    transform: "translate(-50%, -50%)";
+}
