+++ src/index.tsx
@@ -1,18 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { initializeIcons } from "@fluentui/react";
+import { createRoot } from "react-dom/client";
 
 initializeIcons();
 
-ReactDOM.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
-    document.getElementById("root"),
-);
+const container = document.getElementById("root");
+const root = createRoot(container!);
+root.render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
